refactor: 시카, 범블비 리뷰 반영
