+++ peloton-frontend/component/profile/ProfileImageEditButton.js
@@ -1,13 +1,12 @@
 import React from "react";
-import ProfileImage from "../ProfileImage";
-import { Alert, Image, Linking, StyleSheet, TouchableOpacity, View, } from "react-native";
+import { Alert, Linking, TouchableOpacity } from "react-native";
 import { useRecoilState, useRecoilValue } from "recoil";
-import { userInfoState, userTokenState } from "../../atoms";
-import * as ImagePicker from "expo-image-picker";
+import { userInfoState, userTokenState } from "../atoms";
 import Axios from "axios";
-import { SERVER_BASE_URL } from "../../../utils/constants";
+import { SERVER_BASE_URL } from "../../utils/constants";
+import * as ImagePicker from "expo-image-picker";
 
-const ProfileEditImage = () => {
+const ProfileImageEditButton = ({ children }) => {
   const [userInfo, setUserInfo] = useRecoilState(userInfoState);
   const token = useRecoilValue(userTokenState);
 
@@ -16,26 +15,17 @@ const ProfileEditImage = () => {
     formData.append("profile_image", {
       uri: selectedImage,
       type: "image/jpeg",
-      name: `${userInfo.name}-${selectedImage.substring(0, 10)}`,
+      name: selectedImage.substring(9),
     });
     Axios.post(`${SERVER_BASE_URL}/api/members/profile`, formData, {
       headers: {
         "Content-Type": "multipart/form-data",
         Authorization: `Bearer ${token}`,
       },
-    })
-      .then((response) => {
-        setUserInfo({
-          ...userInfo,
-          profile: {
-            baseImageUrl: selectedImage,
-          },
-        });
-      })
-      .catch((err) => console.log(err));
+    }).catch((err) => alert(err));
   };
 
-  const openImagePickerAsync = async () => {
+  const pickAndChangeProfileImage = async () => {
     const permissionResult = await ImagePicker.requestCameraRollPermissionsAsync();
     if (permissionResult.granted === false) {
       Alert.alert(
@@ -48,7 +38,7 @@ const ProfileEditImage = () => {
           },
           { text: "OK", onPress: async () => await Linking.openSettings() },
         ],
-        { cancelable: false }
+        { cancelable: false },
       );
     } else {
       const pickerResult = await ImagePicker.launchImageLibraryAsync({
@@ -62,38 +52,21 @@ const ProfileEditImage = () => {
         return;
       }
       const selectedImage = pickerResult.uri;
+      await setUserInfo({
+        ...userInfo,
+        profile: {
+          baseImageUrl: selectedImage,
+        },
+      });
       await requestChangeImage(selectedImage);
     }
   };
 
   return (
-    <View style={styles.imageContainer}>
-      <ProfileImage image={userInfo.profile.baseImageUrl} />
-      <TouchableOpacity onPress={openImagePickerAsync}>
-        <Image
-          style={styles.profileEditButton}
-          source={require("../../../assets/icn_edit.png")}
-        />
-      </TouchableOpacity>
-    </View>
+    <TouchableOpacity onPress={pickAndChangeProfileImage}>
+      {children}
+    </TouchableOpacity>
   );
 };
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#eceff0",
-  },
-  imageContainer: {
-    alignItems: "center",
-    justifyContent: "center",
-    marginTop: 25,
-  },
-  profileEditButton: {
-    width: 40,
-    height: 40,
-    left: 30,
-    bottom: 30,
-  },
-});
 
-export default ProfileEditImage;
+export default ProfileImageEditButton;
