+++ src/main/java/com/woowacourse/pelotonbackend/infra/upload/S3UploadService.java
@@ -9,6 +9,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
+import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.woowacourse.pelotonbackend.common.exception.UploadFailureException;
 
@@ -30,16 +31,17 @@ public class S3UploadService implements UploadService {
     }
 
     public String uploadImage(final MultipartFile file, final String path) {
-        final String fileName = file.getOriginalFilename();
+        final String key = String.format("%s%s", path, file.getOriginalFilename());
 
         try {
-            final String key = String.format("%s%s", path, fileName);
-            amazonS3.putObject(new PutObjectRequest(bucket, key, file.getInputStream(), null)
+            final ObjectMetadata metadata = new ObjectMetadata();
+            metadata.setContentLength(file.getSize());
+            amazonS3.putObject(new PutObjectRequest(bucket, key, file.getInputStream(), metadata)
                 .withCannedAcl(CannedAccessControlList.PublicRead));
         } catch (Exception e) {
             throw new UploadFailureException();
         }
 
-        return amazonS3.getUrl(bucket, fileName).toString();
+        return amazonS3.getUrl(bucket, key).toString();
     }
 }
